+++ tests/Jroc.Test262.Tests/built-ins/String/prototype/JavaScript/S15.5.4_A2.js
@@ -1,0 +1,9 @@
+// Copyright 2009 the Sputnik authors.  All rights reserved.
+// This code is governed by the BSD license found in the LICENSE file.
+/*---
+info: The String prototype object is itself a String object whose value is an empty string
+es5id: 15.5.4_A2
+description: Checking String.prototype
+---*/
+
+console.log(String.prototype == "");